+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.0] - 2022.02.23
+
+### Added
+
+- Added ability to specify SMTP port number
+
+### Updated
+
+- Updated truemail gemspec
+- Updated gem version
+
 ## [2.6.6] - 2022.02.09
 
 ### Added